refactor: streamline chat API by removing unused code and enhancing response handling
- Remove legacy code and unused interfaces from the chat API
- Simplify the response generation process by integrating new utility functions
- Update import statements to reflect the new structure and improve clarity
- Maintain existing functionality while optimizing the overall codebase
