+++ utils/ai/types.ts
@@ -1,0 +1,43 @@
+export interface ChatCallProps {
+  message: string;
+  conversationId: string;
+  fileContent?: string[];
+}
+
+export interface StreamResponseChunk {
+  type: 'thought' | 'honcho' | 'response' | 'pdf';
+  text: string;
+}
+
+export interface Message {
+  id: string;
+  is_user: boolean;
+  content: string;
+}
+
+export interface MetaMessage {
+  message_id: string;
+  content: string;
+}
+
+export interface UserData {
+  appId: string;
+  userId: string;
+}
+
+export interface ValidationResult {
+  isAuthorized: boolean;
+  error?: string;
+  status?: number;
+  userData?: UserData;
+  supabaseUser?: any;
+}
+
+export interface ConversationHistory {
+  messages: Message[];
+  thoughts: MetaMessage[];
+  honchoMessages: MetaMessage[];
+  pdfMessages: MetaMessage[];
+  summaries: MetaMessage[];
+  collectionId?: string;
+}
